+++ BusinessLogic/DtoModels/RequestDto/EmployeeReqDto.cs
@@ -5,7 +5,7 @@ namespace BusinessLogic.DtoModels.RequestDto
     public class EmployeeReqDto : PersonReqDto
     {
         public UserPosition Position { get; set; }
-        public double? Prize { get; set; }
+        public decimal? Prize { get; set; }
         public string Password { get; set; }
     }
 }
