+++ BusinessLogic/DtoModels/ResponseDto/RentalResDto.cs
@@ -7,8 +7,8 @@ namespace BusinessLogic.DtoModels.ResponseDto
         public string ClientCNumber { get; set; }
         public DateTime DateOfIssue { get; set; }
         public DateTime DateOfRental { get; set; }
-        public double PledgeSum { get; set; }
-        public double? ReturnSum { get; set; }
+        public decimal PledgeSum { get; set; }
+        public decimal? ReturnSum { get; set; }
         public string DiscTitle { get; set; }
         public string EmployeeFName { get; set; }
         public int ProductId { get; set; }
